🚚 api: Move organization route to certification/results context
