+++ api/tests/certification/results/unit/application/organization-controller_test.js
@@ -1,5 +1,5 @@
-import { organizationController } from '../../../../../src/certification/course/application/organization-controller.js';
-import { usecases } from '../../../../../src/certification/course/domain/usecases/index.js';
+import { organizationController } from '../../../../../src/certification/results/application/organization-controller.js';
+import { usecases } from '../../../../../src/certification/results/domain/usecases/index.js';
 import { domainBuilder, expect, hFake, sinon } from '../../../../test-helper.js';
 import { getI18n } from '../../../../tooling/i18n/i18n.js';
